Move type guards to utils
